ProductController 정책 변경

+++ DressHub/src/main/java/com/newoneplus/dresshub/Repository/ProductImageRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface ProductImageRepository extends JpaRepository<ProductImage, Integer> {
-     List<ProductImage> findAllByProductIdOrderByIdDesc(int productId);
+     List<ProductImage> findAllByProductIdOrderByIdDesc(long productId);
      List<ProductImage> findAllByOrderByIdDesc();
 
 }
